Refactor: cts: Move cts.CTSaudits to _cts module.
